+++ libalx/inc/alx_seed.h
@@ -1,6 +1,8 @@
 # ifndef		SEED_H
 	# define	SEED_H
 
-	int	seedf		(int a, int b, int c);
+	#include <stdint.h>
+
+	int64_t	seedf		(int64_t a, int64_t b, int64_t c);
 
 # endif			/* seed.h */
